moved exceptions out of opentrons.hardware_control.types into new file opentrons.hardware_control.errors satisfy mypy
