+++ api/src/opentrons/hardware_control/api.py
@@ -37,7 +37,6 @@ from .module_control import AttachedModulesControl
 from .types import (
     Axis,
     CriticalPoint,
-    MustHomeError,
     DoorState,
     DoorStateNotification,
     ErrorMessageNotification,
@@ -45,6 +44,9 @@ from .types import (
     HardwareAction,
     MotionChecks,
     PauseType,
+)
+from .errors import (
+    MustHomeError,
 )
 from . import modules
 from .robot_calibration import (
