+++ api/src/opentrons/hardware_control/instruments/ot2/pipette_handler.py
@@ -22,11 +22,13 @@ from opentrons import types as top_types
 from opentrons.hardware_control.types import (
     CriticalPoint,
     HardwareAction,
-    TipAttachedError,
-    NoTipAttachedError,
     Axis,
     OT3Axis,
     OT3Mount,
+)
+from opentrons.hardware_control.errors import (
+    TipAttachedError,
+    NoTipAttachedError,
 )
 from opentrons.hardware_control.constants import (
     SHAKE_OFF_TIPS_SPEED,
