+++ api/src/opentrons/hardware_control/instruments/ot3/pipette_handler.py
@@ -20,10 +20,12 @@ from opentrons import types as top_types
 from opentrons.hardware_control.types import (
     CriticalPoint,
     HardwareAction,
-    TipAttachedError,
-    NoTipAttachedError,
     OT3Axis,
     OT3Mount,
+)
+from opentrons.hardware_control.errors import (
+    TipAttachedError,
+    NoTipAttachedError,
 )
 from opentrons.hardware_control.constants import (
     SHAKE_OFF_TIPS_SPEED,
