+++ api/src/opentrons/hardware_control/ot3api.py
@@ -55,7 +55,6 @@ from .util import DeckTransformState
 from .types import (
     Axis,
     CriticalPoint,
-    MustHomeError,
     DoorState,
     DoorStateNotification,
     ErrorMessageNotification,
@@ -70,8 +69,8 @@ from .types import (
     GripperJawState,
     InstrumentProbeType,
     GripperProbe,
-    GripperNotAttachedError,
 )
+from .errors import MustHomeError, GripperNotAttachedError
 from . import modules
 from .robot_calibration import (
     OT3Transforms,
