+++ api/src/opentrons/hardware_control/types.py
@@ -11,18 +11,6 @@ MODULE_LOG = logging.getLogger(__name__)
 MachineType = Literal["ot2", "ot3"]
 
 
-class OutOfBoundsMove(RuntimeError):
-    def __init__(self, message: str):
-        self.message = message
-        super().__init__()
-
-    def __str__(self) -> str:
-        return f"OutOfBoundsMove: {self.message}"
-
-    def __repr__(self) -> str:
-        return f"<{str(self.__class__)}: {self.message}>"
-
-
 class MotionChecks(enum.Enum):
     NONE = 0
     LOW = 1
@@ -467,22 +455,6 @@ class AionotifyEvent:
         return cls(flags=Flag, name=name)
 
 
-class ExecutionCancelledError(RuntimeError):
-    pass
-
-
-class MustHomeError(RuntimeError):
-    pass
-
-
-class NoTipAttachedError(RuntimeError):
-    pass
-
-
-class TipAttachedError(RuntimeError):
-    pass
-
-
 class GripperJawState(enum.Enum):
     UNHOMED = enum.auto()
     #: the gripper must be homed before it can do anything
@@ -512,42 +484,3 @@ class GripperProbe(enum.Enum):
             return InstrumentProbeType.PRIMARY
         else:
             return InstrumentProbeType.SECONDARY
-
-
-class InvalidMoveError(ValueError):
-    pass
-
-
-class GripperNotAttachedError(Exception):
-    """An error raised if a gripper is accessed that is not attached"""
-
-    pass
-
-
-class InvalidPipetteName(KeyError):
-    """Raised for an invalid pipette."""
-
-    def __init__(self, name: int, mount: OT3Mount) -> None:
-        self.name = name
-        self.mount = mount
-
-    def __repr__(self) -> str:
-        return f"<{self.__class__.__name__}: name={self.name} mount={self.mount}>"
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}: Pipette name key {self.name} on mount {self.mount.name} is not valid"
-
-
-class InvalidPipetteModel(KeyError):
-    """Raised for a pipette with an unknown model."""
-
-    def __init__(self, name: str, model: str, mount: OT3Mount) -> None:
-        self.name = name
-        self.model = model
-        self.mount = mount
-
-    def __repr__(self) -> str:
-        return f"<{self.__class__.__name__}: name={self.name}, model={self.model}, mount={self.mount}>"
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}: {self.name} on {self.mount.name} has an unknown model {self.model}"
