+++ api/tests/opentrons/hardware_control/test_moves.py
@@ -13,10 +13,12 @@ from opentrons.calibration_storage.types import (
 from opentrons.hardware_control.types import (
     Axis,
     CriticalPoint,
-    OutOfBoundsMove,
     MotionChecks,
+)
+from opentrons.hardware_control.errors import (
     MustHomeError,
     InvalidMoveError,
+    OutOfBoundsMove,
 )
 from opentrons.hardware_control.robot_calibration import (
     RobotCalibration,
