+++ api/tests/opentrons/hardware_control/test_ot3_api.py
@@ -21,11 +21,13 @@ from opentrons.hardware_control.instruments.ot3.pipette_handler import (
 from opentrons.hardware_control.types import (
     OT3Mount,
     OT3Axis,
-    GripperNotAttachedError,
-    InvalidMoveError,
     CriticalPoint,
     GripperProbe,
     InstrumentProbeType,
+)
+from opentrons.hardware_control.errors import (
+    GripperNotAttachedError,
+    InvalidMoveError,
 )
 from opentrons.hardware_control.ot3api import OT3API
 from opentrons.hardware_control import ThreadManager
