+++ api/tests/opentrons/protocol_api_old/core/simulator/test_instrument_context.py
@@ -4,8 +4,10 @@ from typing import Callable
 import pytest
 from pytest_lazyfixture import lazy_fixture  # type: ignore[import]
 
-from opentrons.hardware_control import NoTipAttachedError
-from opentrons.hardware_control.types import TipAttachedError
+from opentrons.hardware_control import (
+    NoTipAttachedError,
+    TipAttachedError,
+)
 from opentrons.protocol_api.core.common import InstrumentCore, LabwareCore
 from opentrons.types import Location, Point
